+++ tests/test_metadata.py
@@ -10,7 +10,7 @@ import pytest
 import pytz
 import pandas as pd
 import hatyan
-from hatyan.metadata import metadata_from_obj, metadata_compare, wns_from_metadata #, metadata_from_diablocks, metadata_add_to_obj
+from hatyan.metadata import metadata_from_obj, metadata_compare, wns_from_metadata
 
 dir_tests = os.path.dirname(__file__) #F9 doesnt work, only F5 (F5 also only method to reload external definition scripts)
 dir_testdata = os.path.join(dir_tests,'data_unitsystemtests')
@@ -34,10 +34,10 @@ def test_readts_dia_metadata_multifile():
      'timestep_unit': 'min',
      'TYP': 'TE',
      'groepering': 'NVT',
-     'tzone': pytz.FixedOffset(60),
      'origin': 'from timeseries dia file'}
     
     assert meta_fromts == meta_expected
+    assert ts_measurements_group0.index.tz == pytz.FixedOffset(60)
 
 
 @pytest.mark.unittest
@@ -67,7 +67,6 @@ def test_anapred_metadata():
      'timestep_unit': 'min',
      'TYP': 'TE',
      'groepering': 'NVT',
-     'tzone': pytz.FixedOffset(60),
      'origin': 'from timeseries dia file',
      'nodalfactors': True,
      'xfac': False,
@@ -79,6 +78,8 @@ def test_anapred_metadata():
     
     assert meta_fromts_xfac0 == meta_expected_xfac0
     assert meta_fromts_xfac1 == meta_expected_xfac1
+    assert pred_xfac0.index.tz == pytz.FixedOffset(60)
+    assert pred_xfac1.index.tz == pytz.FixedOffset(60)
 
 
 @pytest.mark.unittest
@@ -108,7 +109,6 @@ def test_hwlw_metadata():
      'timestep_unit': 'min',
      'TYP': 'TE',
      'groepering': 'NVT',
-     'tzone': pytz.FixedOffset(60),
      'origin': 'from timeseries dia file'}
     
     pred_ext_meta_expected = {'station': 'VLISSGN',
@@ -121,7 +121,6 @@ def test_hwlw_metadata():
      'timestep_unit': 'min',
      'TYP': 'TE',
      'groepering': 'NVT',
-     'tzone': pytz.FixedOffset(60),
      'origin': 'from timeseries dia file',
      'nodalfactors': True,
      'xfac': False,
@@ -130,6 +129,10 @@ def test_hwlw_metadata():
     
     assert pred_ext_meta == pred_ext_meta_expected
     assert meas_ext_meta == meas_ext_meta_expected
+    assert ts_measurements_group0.index.tz == pytz.FixedOffset(60)
+    assert meas_ext.index.tz == pytz.FixedOffset(60)
+    assert pred.index.tz == pytz.FixedOffset(60)
+    assert pred_ext.index.tz == pytz.FixedOffset(60)
 
 
 @pytest.mark.unittest
@@ -148,10 +151,10 @@ def test_readts_dia_metadata_multiblock():
      'timestep_unit': None,
      'TYP': 'TN',
      'groepering': 'GETETM2',
-     'tzone': pytz.FixedOffset(60),
      'origin': 'from timeseries dia file'}
     
     assert meta_fromts == meta_expected
+    assert ts_measurements_group0.index.tz == pytz.FixedOffset(60)
 
 
 @pytest.mark.unittest